Consolidate Linode services.
