+++ src/features/linodes/LinodesDetail/LinodeRebuild/LinodeRebuild.tsx
@@ -17,7 +17,7 @@ import MenuItem from 'material-ui/Menu/MenuItem';
 
 import { sendToast } from 'src/features/ToastNotifications/toasts';
 import { getImages } from 'src/services/images';
-import { rebuild as rebuildLinode } from 'src/services/linodes';
+import { rebuildLinode } from 'src/services/linodes';
 import { resetEventsPolling } from 'src/events';
 import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
 import Select from 'src/components/Select';
