--- src/service/linodes.ts
@@ -1,6 +1,0 @@
-import Axios from 'axios';
-import { API_ROOT } from 'src/constants';
-
-export const resize = (id: number, type: string): Promise<{}> => Axios
-  .post(`${API_ROOT}/linode/instances/${id}/resize`, { type });
-
